requirements.txt
jupyter
ipywidgets
ipykernel
open-clip-torch
torch
torchvision
torchaudio
fastapi[standard]
python-multipart
numpy<2